+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import "./index.css";
 import "./styles/index.scss";
-import Home from "./pages/Home.js";
-import Dev from "./pages/Dev.js";
+import Home from "./pages/Home";
+import Dev from "./pages/Dev";
 import Navbar from "./componentes/Navbar/Navbar";
 import Footer from "./componentes/Footer/Footer";
 import { WarningHeader } from "./componentes/WarningHeader/WarningHeader";
