+++ src/assets/iconos/BreakPoints.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 /**
  * @param {string} color - color asignado al icono
  * @param {JSX.Element} svg - SVG element o React-Icon
