+++ src/componentes/Cake-sum/Cake-sum.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 
 const CakeSum = ({ limite, className }) => {
   const [suma, setSuma] = useState(limite);
