+++ src/componentes/Dropdown/Dropdown.jsx
@@ -1,5 +1,5 @@
 import Link from "../Link/Link";
-import React, { useState } from "react";
+import { useState } from "react";
 import DropdownItem from "./DropdownItems/DropdownItems.jsx";
 import "./Dropdown.scss";
