+++ src/componentes/Icono/PancakeSwapIcono.jsx
@@ -1,8 +1,7 @@
-import React from "react";
-import asset from "../../assets/conejo.svg"
+import asset from "../../assets/conejo.svg";
 
 const PancakeSwapIcono = () => {
-  return <img src={asset} alt="conejo" />
+  return <img src={asset} alt="conejo" />;
 };
 
 export default PancakeSwapIcono;
