+++ src/componentes/Puntos/Puntos.jsx
