+++ src/componentes/Section/BasePage.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const BasePage = ({ children }) => {
   return <div className="base-page">{children}</div>;
 };
