+++ src/componentes/Section/Section.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Section.scss";
 
 export const MainSection = ({ children }) => {
