+++ src/componentes/Section/Section2/Box.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import classNames from "classnames";
 
 export const Box = ({
