--- src/vite.config.js
@@ -1,12 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react-swc";
-import sassDts from "vite-plugin-sass-dts";
-import sass from "vite-plugin-sass";
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react(), sassDts(), sass()],
-  server: {
-    port: 8080,
-  },
-});
